Revert "New"
This reverts commit a4005b9c0c0837919c33768c5bd8f853927576ad.

+++ Hardis'sApplication-ejb/src/java/entitee/PriseEnCharge.java
@@ -26,7 +26,7 @@ public class PriseEnCharge implements Serializable {
 
     private double PrixRemboursement;
 
-    private double BaseRemboursement;
+    private int BaseRemboursement;
 
     private Unite Unite;
 
@@ -60,11 +60,11 @@ public class PriseEnCharge implements Serializable {
         this.Unite = Unite;
     }
 
-    public double getBaseRemboursement() {
+    public int getBaseRemboursement() {
         return BaseRemboursement;
     }
 
-    public void setBaseRemboursement(double BaseRemboursement) {
+    public void setBaseRemboursement(int BaseRemboursement) {
         this.BaseRemboursement = BaseRemboursement;
     }
